backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
supabase>=1.0.0
python-dotenv>=1.0.0
langchain>=0.0.400
openai>=1.0.0
pydantic>=1.10.0
httpx>=0.24.0
redis>=4.6.0
rq>=1.1.0